# Add optimizations for meta records of which the underlying expressions use equal operator