This PR is based on #1541, it does not make sense to review it before that one is merged.

When we filter by meta records of which the underlying expressions all use the equal operator we can save a lot of tag index lookup by building a set of acceptable tags based on the meta record expressions and then simple filtering by them, instead of having to evaluate each of them by doing the tag lookup on the tag index. 

These 3 benchmarks are testing cases where we filter by meta tags which have a large number of underlying expressions which all use equal operators. For example a datacenter mapping to a list of `host=X` tag/value pairs.

```
benchmark                                                                                  old ns/op        new ns/op       delta
BenchmarkFilter100kByMetaTagWithIndexSize1mAnd50kMetaRecords-12                            95441374800      285884142       -99.70%
BenchmarkFilter10kByMetaTagWithIndexSize1mAnd10kMetaRecords-12                             2008162900       30703885        -98.47%
BenchmarkFilter100kByMetaTagWithIndexSize1mAnd50kMetaRecordsWithMultipleExpressions-12     115969750200     12734048100     -89.02%
```

grafana/metrictank-ops/issues/524