@Dieterbe 2019-11-20
is there any types of queries that might get a slow down because of this - or the other linked to - PR?