@replay 2019-11-20
In the case of this PR #1542 I'm quite confident that it won't slow down any other queries. It detects a specific situation, which is that all expressions underlying a meta tag which is used to filter down the result set are using the `=` operator, and then it applies an optimization which is very specific to this situation. Since this is a very common use case of which we know that big customers are using it often it makes sense to add an optimization which is very specific to that use case.

In the case of #1541 it is possible that certain queries get slowed down if for example a meta tag is very cheap to evaluate because it only uses simple operators (such as `=`) and only matches a small number of metrics, while at the same time another expression of the same query is operating on the metric tag index while using expensive operators (such as `=~`) and matching a large number of series. 
Based on what we know the common use case of the meta tag feature is, this is an unlikely use case though.